Handle `React.ReactChild` correctly
