+++ packages/uxpin-merge-cli/package.json
@@ -133,7 +133,7 @@
     "sortobject": "^1.1.1",
     "source-map-support": "^0.4.18",
     "tslib": "^1.9.0",
-    "typescript": "4.4.2",
+    "typescript": "4.8.4",
     "uuid": "^3.3.2",
     "webpack-merge": "^4.1.2",
     "webpack-virtual-modules": "^0.5.0"
