--- packages/uxpin-merge-cli/src/steps/serialization/component/implementation/typescript/property/type/checker/isBooleanLike.ts
@@ -1,6 +1,0 @@
-import * as ts from 'typescript';
-
-export function isBooleanLike(type: ts.Type): boolean {
-  // tslint:disable-next-line no-bitwise
-  return Boolean(type.flags & ts.TypeFlags.BooleanLike);
-}
