--- packages/uxpin-merge-cli/src/steps/serialization/component/implementation/typescript/property/type/checker/isCallable.ts
@@ -1,6 +1,0 @@
-import * as ts from 'typescript';
-import { isObjectLike } from './isObjectLike';
-
-export function isCallable(type: ts.Type): boolean {
-  return isObjectLike(type) && !!type.getCallSignatures().length;
-}
