--- packages/uxpin-merge-cli/src/steps/serialization/component/implementation/typescript/property/type/checker/isEnum.ts
@@ -1,9 +1,0 @@
-import * as ts from 'typescript';
-
-export function isEnum(type: ts.Type): type is ts.LiteralType {
-  if (type.symbol?.valueDeclaration) {
-    return ts.isEnumMember(type.symbol.valueDeclaration);
-  }
-
-  return false;
-}
