--- packages/uxpin-merge-cli/src/steps/serialization/component/implementation/typescript/property/type/checker/isIntersection.ts
@@ -1,5 +1,0 @@
-import * as ts from 'typescript';
-
-export function isIntersection(type: ts.Type): type is ts.IntersectionType {
-  return type.isIntersection();
-}
