--- packages/uxpin-merge-cli/src/steps/serialization/component/implementation/typescript/property/type/checker/isIntersectionOfObjects.ts
@@ -1,7 +1,0 @@
-import * as ts from 'typescript';
-import { isIntersection } from './isIntersection';
-import { isObjectLike } from './isObjectLike';
-
-export function isIntersectionOfObjects(type: ts.Type): boolean {
-  return isIntersection(type) && type.types.every(isObjectLike);
-}
