--- packages/uxpin-merge-cli/src/steps/serialization/component/implementation/typescript/property/type/checker/isLiteral.ts
@@ -1,5 +1,0 @@
-import * as ts from 'typescript';
-
-export function isLiteral(type: ts.Type): type is ts.LiteralType {
-  return type.isLiteral();
-}
