--- packages/uxpin-merge-cli/src/steps/serialization/component/implementation/typescript/property/type/checker/isObjectLike.ts
@@ -1,7 +1,0 @@
-import * as ts from 'typescript';
-import { isIntersectionOfObjects } from './isIntersectionOfObjects';
-import { isObject } from './isObject';
-
-export function isObjectLike(type: ts.Type): boolean {
-  return isObject(type) || isIntersectionOfObjects(type);
-}
