--- packages/uxpin-merge-cli/src/steps/serialization/component/implementation/typescript/property/type/checker/isUnion.ts
@@ -1,5 +1,0 @@
-import * as ts from 'typescript';
-
-export function isUnion(type: ts.Type): type is ts.UnionType {
-  return type.isUnion();
-}
